# front+back_renew/docs/api/openapi.yaml
openapi: 3.1.0
info:
  title: AI Business Validation Platform Document API
  version: 0.10.0-phase10
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    Canonical authentication, document upload, asynchronous processing,
    structured-plan mutation, and explicit confirmation contract.
servers:
  - url: /api/v1
...
security:
  - platformAuth: []
paths:
  /auth/signup:
    post:
      tags: [Authentication]
      summary: Register an active user and issue an access/refresh pair
      operationId: signup
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/SignupRequest'}
      responses:
        '201':
          description: User registered
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '409': {$ref: '#/components/responses/StandardError'}
  /auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate with normalized email and password
      operationId: login
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/LoginRequest'}
      responses:
        '200':
          description: Authentication succeeded
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AuthApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /auth/refresh:
    post:
      tags: [Authentication]
      summary: Rotate a valid refresh token and issue a new pair
      operationId: refreshAccessToken
      x-implementation-phase: Phase 3
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Token pair rotated
          content:
            application/json:
              schema: {$ref: '#/components/schemas/TokenPairApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /auth/logout:
    post:
      tags: [Authentication]
      summary: Revoke the supplied current refresh token
      operationId: logout
      x-implementation-phase: Phase 3
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/RefreshRequest'}
      responses:
        '200':
          description: Refresh token revoked
          content:
            application/json:
              schema: {$ref: '#/components/schemas/EmptyApiResponse'}
        '400': {$ref: '#/components/responses/StandardError'}
        '401': {$ref: '#/components/responses/StandardError'}
  /users/me:
    get:
      tags: [Users]
      summary: Get the current authenticated user
      operationId: getCurrentUser
      x-implementation-phase: Phase 3
      responses:
        '200':
          description: Current user
          content:
            application/json:
              schema: {$ref: '#/components/schemas/UserApiResponse'}
        '401': {$ref: '#/components/responses/StandardError'}
...
components:
  securitySchemes:
    platformAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: >
        Phase 3 access JWT. The development X-User-Id adapter is profile-bound
        and is deliberately excluded from the public contract.
...
  responses:
    StandardError:
      description: Standard API error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorApiResponse'
  schemas:
    Meta:
      type: object
      required: [requestId, timestamp]
      properties:
        requestId: {type: string}
        timestamp: {type: string, format: date-time}
    ApiError:
      type: object
      required: [code, message, fieldErrors, retryable]
      properties:
        code: {type: string}
        message: {type: string}
        fieldErrors:
          type: array
          items:
            type: object
            required: [field, message]
            properties:
              field: {type: string}
              message: {type: string}
        retryable: {type: boolean}
    ErrorApiResponse:
      type: object
      required: [success, error, meta]
      properties:
        success: {type: boolean, const: false}
        error: {$ref: '#/components/schemas/ApiError'}
        meta: {$ref: '#/components/schemas/Meta'}
    SignupRequest:
      type: object
      required: [email, password, displayName]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password:
          type: string
          format: password
          minLength: 8
          description: Maximum 72 UTF-8 bytes; never persisted or logged raw.
        displayName: {type: string, minLength: 1, maxLength: 50}
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email: {type: string, format: email, maxLength: 254}
        password: {type: string, format: password}
    RefreshRequest:
      type: object
      required: [refreshToken]
      properties:
        refreshToken:
          type: string
          minLength: 1
          description: Raw token is accepted only at this boundary and stored as SHA-256.
    User:
      type: object
      required: [id, email, displayName, role, status]
      properties:
        id: {type: integer, format: int64}
        email: {type: string, format: email}
        displayName: {type: string}
        role: {type: string, enum: [USER, ADMIN]}
        status:
          type: string
          enum: [PENDING, ACTIVE, LOCKED, SUSPENDED, WITHDRAWN]
    TokenPair:
      type: object
      required:
        [tokenType, accessToken, accessTokenExpiresIn, refreshToken,
         refreshTokenExpiresIn]
      properties:
        tokenType: {type: string, const: Bearer}
        accessToken: {type: string}
        accessTokenExpiresIn: {type: integer, format: int64, minimum: 1}
        refreshToken: {type: string}
        refreshTokenExpiresIn: {type: integer, format: int64, minimum: 1}
    Auth:
      type: object
      required: [user, tokens]
      properties:
        user: {$ref: '#/components/schemas/User'}
        tokens: {$ref: '#/components/schemas/TokenPair'}
...
    AuthApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/Auth'}
        meta: {$ref: '#/components/schemas/Meta'}
...
    TokenPairApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/TokenPair'}
        meta: {$ref: '#/components/schemas/Meta'}
    UserApiResponse:
      type: object
      required: [success, data, meta]
      properties:
        success: {type: boolean, const: true}
        data: {$ref: '#/components/schemas/User'}
        meta: {$ref: '#/components/schemas/Meta'}
    EmptyApiResponse:
      type: object
      required: [success, meta]
      properties:
        success: {type: boolean, const: true}
        meta: {$ref: '#/components/schemas/Meta'}
...